[TS migration] Migrate 'mocks' workflow test to TypeScript
