--- workflow_tests/mocks/reviewerChecklistMocks.js
@@ -1,9 +1,0 @@
-const utils = require('../utils/utils');
-
-// checklist
-const REVIEWERCHECKLIST__CHECKLIST__REVIEWERCHECKLIST_JS__STEP_MOCK = utils.createMockStep('reviewerChecklist.js', 'reviewerChecklist.js', 'CHECKLIST', ['GITHUB_TOKEN'], []);
-const REVIEWERCHECKLIST__CHECKLIST__STEP_MOCKS = [REVIEWERCHECKLIST__CHECKLIST__REVIEWERCHECKLIST_JS__STEP_MOCK];
-
-module.exports = {
-    REVIEWERCHECKLIST__CHECKLIST__STEP_MOCKS,
-};
